moved forceselection to a utils file
